up to next version

+++ src/backend/msc-handler.cpp
@@ -238,7 +238,7 @@ void	mscHandler::stop_service	(descriptorType *d, int flag) {
 
 void	mscHandler::stop_service	(int subchId, int flag) {
 	locker. lock ();
-	for (int i = 0; i < theBackends. size (); i ++) {
+	for (int i = 0; i < (int)(theBackends. size ()); i ++) {
 	   Backend *b = theBackends. at (i);
 	   if ((b -> subChId == subchId) && (b -> borf == flag)) {
 	      fprintf (stderr, "stopping subchannel %d\n", subchId);
@@ -274,7 +274,7 @@ bool	mscHandler::set_Channel (descriptorType *d,
 	                                     dump,
 	                                     flag));
 	fprintf (stderr, "we have now %d backends running\n",
-	                        theBackends. size ());
+	                        (int)(theBackends. size ()));
 	return true;
 }
